Moved open attachment method to separate controller
