+++ build.gradle
@@ -178,8 +178,8 @@ project(':geoportaal-harvester') {
 		environment 'DB_USER', dbUser
 		environment 'DB_PASSWORD', dbPassword
 		
-		environment 'DATA_URL', 'http://test-metadata.geodataoverijssel.nl/metadata/service/'
-		environment 'DATA_NAME', 'service'
+		environment 'DATA_URL', 'https://test-admin.geoportaaloverijssel.nl/metadata/dc/'
+		environment 'DATA_NAME', 'dc'
 		environment 'DATA_SECURED', 'false'
 		environment 'HARVEST_INTERVAL', '30'
 	}
