+++ geoportaal-beheer/conf/routes
@@ -19,7 +19,7 @@ GET		/editStatus						controllers.Index.changeStatus()
 GET		/editSupplier					controllers.Index.changeSupplier()
 GET		/cancel							controllers.Metadata.cancel(text ?= "", supplier ?= "none", status ?= "none", format ?= "none", dateUpdateStart ?= "", dateUpdateEnd ?= "")
 GET		/delete							controllers.Index.deleteMetadata()
-GET		/openAttachment/:attachmentName		controllers.Metadata.openAttachment(attachmentName: String, uuid: String)
+GET		/openAttachment/:attachmentName		controllers.Attachment.openAttachment(attachmentName: String, uuid: String)
 POST	/search							controllers.Index.search()
 POST	/sort							controllers.Index.sort()
 GET		/help							controllers.Help.renderHelp()
